- Optimizations - Windows now automatically centers after update check

+++ pages/js/pairs.js
@@ -2,51 +2,15 @@ const { ipcRenderer } = require('electron');
 const {dialog} = require('electron').remote;
 
 window.onload = function() {
-  if(window.localStorage.getItem("airports") == null){
-    window.localStorage.setItem("airports", JSON.stringify(['icao,iata,name,location,country,timezone,hub,lat,lon,ground_handling_cost,fuel_100ll_cost,fuel_jeta_cost,fuel_mogas_cost,notes'.split(",")]));
-  }
-  if(window.localStorage.getItem("flights") == null){
-    window.localStorage.setItem("flights", JSON.stringify(['airline,flight_number,route_code,callsign,route_leg,dpt_airport,arr_airport,alt_airport,days,dpt_time,arr_time,level,distance,flight_time,flight_type,load_factor,load_factor_variance,pilot_pay,route,notes,start_date,end_date,active,subfleets,fares,fields'.split(",")]));
-  }
-  if(window.localStorage.getItem("pairs") == null){
-    window.localStorage.setItem("pairs", JSON.stringify(['dep_icao,arr_icao'.split(",")]));
-  }
-
   let data = JSON.parse(window.localStorage.getItem("airports"));
   if(data.length <= 1){
     alert("Please add an airport first!");
     window.location.href = "airports.html";
   }
-
   load(JSON.parse(window.localStorage.getItem("pairs")));
 }
 
 
-function alert(message) {
-    const options = {
-      type: 'info',
-      buttons: ['OK'],
-      defaultId: 0,
-      title: 'Alert',
-      message: message == undefined ? 'Oops! This is not supposed to be sent! Please report this bug!' : message
-    };
-  
-    dialog.showMessageBox(null, options, (response, checkboxChecked) => {
-    });
-  }
-  
-  function confirm(message) {
-    const options = {
-      type: 'info',
-      buttons: ['Yes', 'No'],
-      defaultId: 1,
-      title: 'Alert',
-      message: message == undefined ? 'Oops! This is not supposed to be sent! Please report this bug!' : message
-    };
-    let msg = dialog.showMessageBox(null, options)
-    return msg
-  }
-  
 
   function createPair(){
     let dep = ipcRenderer.sendSync('editPairs', 0, "DepartureICAO(rqd)", "", true, "");
@@ -61,6 +25,8 @@ function alert(message) {
     if(airportExist(arr) == false)return alert("Arrival Airport not found!");
     if(pairExist(dep, arr) == true)return alert("Pair Already Exists!");
 
+    let warn = confirm("Are you sure you want to add this pair? Added pairs cannot be deleted!");
+    if(warn == 1)return;
     let data = JSON.parse(window.localStorage.getItem("pairs"));
     data.push([dep.toUpperCase(), arr.toUpperCase()]);
     window.localStorage.setItem("pairs", JSON.stringify(data));
